+++ tests/unit/eyes_core/test_check_settings.py
@@ -1,6 +1,7 @@
 import pytest
 from mock import Mock
 
+from applitools.common import MatchLevel
 from applitools.common.accessibility import AccessibilityRegionType
 from applitools.common.geometry import AccessibilityRegion, Region
 from applitools.core import CheckSettings
@@ -44,6 +45,26 @@ def test_give_incorrect_parameters_to_match_regions():
     with pytest.raises(TypeError) as exc_info:
         cs = CheckSettings().floating(1)
         assert exc_info.type == TypeError
+
+
+def test_set_match_regions_level():
+    cs = CheckSettings().layout(Region.EMPTY())
+    assert cs.values.match_level is None
+    cs = cs.layout()
+    assert cs.values.match_level == MatchLevel.LAYOUT
+
+    cs = CheckSettings().content(Region.EMPTY())
+    assert cs.values.match_level is None
+    cs = cs.content()
+    assert cs.values.match_level == MatchLevel.CONTENT
+
+    cs = CheckSettings().strict(Region.EMPTY())
+    assert cs.values.match_level is None
+    cs = cs.strict()
+    assert cs.values.match_level == MatchLevel.STRICT
+
+    cs = CheckSettings().exact()
+    assert cs.values.match_level == MatchLevel.EXACT
 
 
 def test_set_get_accessibility_regions():